refactor: rename tools -> aec
