+++ MANIFEST.in
@@ -1,3 +1,3 @@
 # needed by setup.py
 include requirements.*
-graft tools/config-example
+graft aec/config-example
