+++ aec/config-example/userdata/amzn-install-docker.yaml
