+++ setup.py
@@ -15,7 +15,7 @@ setup(
     long_description_content_type="text/markdown",
     url="https://github.com/seek-oss/aec",
     license="MIT",
-    entry_points={"console_scripts": ["aec = tools.main:main"]},
+    entry_points={"console_scripts": ["aec = aec.main:main"]},
     python_requires=">=3.7",
     packages=find_packages(exclude=["tests"]),
     include_package_data=True,
